Extract `StreamFieldInterface` to `interfaces.py`
